Move PlayerState to own file
